+++ packages/adminui-vue/components/UTableEntity/filter-templates/many/isNull.vue
@@ -9,14 +9,15 @@ export default {
   components: {
     FilterTemplate: require('../../components/FilterTemplate.vue').default
   },
-  
+
   mixins: [require('../mixinForFilter.js')],
 
   methods: {
     getCondition () {
+      const { $ut } = this
       return {
         whereList: [{ condition: 'isNull' }],
-        description: 'isNull'
+        description: $ut('isNull')
       }
     }
   }